+++ Assets/Python/BUG/BugData.py
@@ -37,7 +37,10 @@
 from CvPythonExtensions import CyGame
 import BugUtil
 import cPickle as pickle
+from StoredData import data
 
+
+STORED_DATA = "doc_stored_data"
 
 ## Data Access
 
@@ -100,9 +103,11 @@ def onGameStart(argsList):
 
 def onGameLoad(argsList):
     initGameData().load()
+    data.update(getTable(STORED_DATA).data)
 
 
 def save():
+    getTable(STORED_DATA).setData(data.__dict__)
     getGameData().save()
