+++ Assets/Python/BUG/BugEventManager.py
@@ -103,8 +103,11 @@ import CvEventManager
 import BugCore
 import BugData
 import BugUtil
+import Init
+import Modifiers
 import InputUtil
 import types
+
 
 # BUG - Mac Support - start
 BugUtil.fixSets(globals())
@@ -430,6 +433,8 @@ class BugEventManager(CvEventManager.CvEventManager):
         import BugInit
 
         BugInit.init()
+        Init.init()
+        Modifiers.setup()
         self._handleDefaultEvent(eventType, argsList)
 
     def resetActiveTurn(self, argsList=None):
