+++ Assets/Python/Modifiers.py
@@ -27,12 +27,10 @@ from MiscData import (
 )
 from TimelineData import TIMELINE_TECH_MODIFIER, DateTurn
 from LocationsData import CITIES
-from Events import handler
 
 gc = CyGlobalContext()
 
 
-@handler("GameStart")
 def setup():
     set_modifiers()
     set_diplomacy_modifier()
